- UI done. Need refactor background.
