+++ app/src/main/java/com/example/blanball/presentation/views/screens/onboarding/fillingouttheprofile/FillingOutTheUserProfileScreenStep3.kt
@@ -1,4 +1,4 @@
-package com.example.blanball.presentation.views.screens.onboarding
+package com.example.blanball.presentation.views.screens.onboarding.fillingouttheprofile
 
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.background
@@ -41,18 +41,17 @@ import com.example.blanball.presentation.theme.primaryDark
 import com.example.blanball.presentation.theme.shapes
 import com.example.blanball.presentation.theme.typography
 import com.example.blanball.presentation.views.components.cards.AnimatedPaddingCard
-import com.example.blanball.presentation.views.components.textinputs.BottomLineDefaultTextInput
-import com.example.blanball.utils.ext.isNotValidBirthDay
-import com.example.blanball.utils.ext.isNotValidBirthMonth
-import com.example.blanball.utils.ext.isNotValidBirthYear
-import com.example.blanball.utils.ext.isValidBirthDay
-import com.example.blanball.utils.ext.isValidBirthMonth
-import com.example.blanball.utils.ext.isValidBirthYear
+import com.example.blanball.presentation.views.components.dropdownmenu.CustomDropDownMenu
+import com.example.blanball.presentation.views.components.textinputs.DefaultTextInput
+import com.example.blanball.utils.ext.isNotValidHeight
+import com.example.blanball.utils.ext.isNotValidWeight
+import com.example.blanball.utils.ext.isValidHeight
+import com.example.blanball.utils.ext.isValidWeight
 
 @Composable
-fun FillingOutTheUserProfileScreenStep1(
+fun FillingOutTheUserProfileScreenStep3(
     state: UiState,
-    onFillingOutTheUserProfileStep2Clicked: () -> Unit,
+    onFillingOutTheUserProfileStep4Clicked: () -> Unit,
     onTurnBackClicked: () -> Unit,
 ) {
     val localFocusManager = LocalFocusManager.current
@@ -64,7 +63,7 @@ fun FillingOutTheUserProfileScreenStep1(
     ) {
         (state as? OnboardingScreensStatesMainContract.State)?.let {
             Image(
-                painter = painterResource(id = R.drawable.onboard_bg),
+                painter = painterResource(id = R.drawable.onboard_bg_3),
                 contentDescription = null,
                 modifier = Modifier.fillMaxWidth()
             )
@@ -80,7 +79,7 @@ fun FillingOutTheUserProfileScreenStep1(
                         .verticalScroll(rememberScrollState()),
                 ) {
                     Text(
-                        text = stringResource(R.string.lets_meet),
+                        text = stringResource(R.string.sports_characteristics),
                         modifier = Modifier.fillMaxWidth(),
                         style = typography.h2,
                         color = primaryDark,
@@ -91,93 +90,139 @@ fun FillingOutTheUserProfileScreenStep1(
                             .padding(top = 20.dp)
                             .align(Alignment.CenterHorizontally)
                     ) {
-                        Image(
-                            painter = painterResource(R.drawable.stepline_1),
-                            contentDescription = null,
-                            Modifier.weight(1f)
-                        )
                         repeat(3) {
-                            Spacer(modifier = Modifier.size(2.dp))
                             Image(
-                                painter = painterResource(id = R.drawable.empty_stepline),
+                                painter = painterResource(R.drawable.stepline_1),
                                 contentDescription = null,
                                 Modifier.weight(1f)
                             )
+                            Spacer(modifier = Modifier.size(2.dp))
                         }
+                        Image(
+                            painter = painterResource(id = R.drawable.empty_stepline),
+                            contentDescription = null,
+                            Modifier.weight(1f)
+                        )
                     }
                     Spacer(modifier = Modifier.size(24.dp))
                     Text(
-                        text = stringResource(id = R.string.when_were_you_born),
+                        text = stringResource(id = R.string.physical_parameters),
                         style = typography.h5,
                         color = primaryDark
                     )
                     Row(horizontalArrangement = Arrangement.spacedBy(8.dp)) {
-                        BottomLineDefaultTextInput(
-                            labelResId = R.string.day,
-                            modifier = Modifier.weight(1f),
-                            value = it.dayBirthdayState.value,
-                            onValueChange = { state.dayBirthdayState.value = it },
+                        DefaultTextInput(
+                            modifier = Modifier
+                                .weight(0.5f)
+                                .fillMaxWidth(),
                             state = it,
+                            labelResId = R.string.height,
+                            value = it.heightState.value,
+                            onValueChange = { state.heightState.value = it },
                             transformation = VisualTransformation.None,
                             keyboardOptions = KeyboardOptions.Default.copy(imeAction = ImeAction.Next, keyboardType = KeyboardType.Number),
                             isError = when {
                                 it.isErrorRequestToFinishOutTheProfile.value -> true
-                                it.dayBirthdayState.value.isNotValidBirthDay() -> true
+                                it.heightState.value.isNotValidHeight() ->  true
                                 else -> false
                             },
                             errorMessage = when {
                                 it.isErrorRequestToFinishOutTheProfile.value -> stringResource(id = R.string.invalid_credential_error)
-                                it.dayBirthdayState.value.isNotValidBirthDay() -> stringResource(id = R.string.birth_day_valid_error)
+                                it.heightState.value.isNotValidHeight() ->  stringResource(id = R.string.height_valid_error)
                                 else -> {("")}
                             }
                         )
-                        BottomLineDefaultTextInput(
-                            labelResId = R.string.month,
-                            value = it.monthBirthdayState.value,
-                            modifier = Modifier.weight(1f),
+                        DefaultTextInput(
+                            modifier = Modifier
+                                .weight(0.5f)
+                                .fillMaxWidth(),
                             state = it,
+                            labelResId = R.string.weight,
+                            value = it.weightState.value,
+                            onValueChange = { state.weightState.value = it },
                             transformation = VisualTransformation.None,
-                            onValueChange = { state.monthBirthdayState.value = it },
-                            keyboardOptions = KeyboardOptions.Default.copy(imeAction = ImeAction.Next, keyboardType = KeyboardType.Number),
-                            isError = when {
-                                it.isErrorRequestToFinishOutTheProfile.value -> true
-                                it.monthBirthdayState.value.isNotValidBirthMonth() -> true
-                                else -> false
-                            },
-                            errorMessage = when {
-                                it.isErrorRequestToFinishOutTheProfile.value -> stringResource(id = R.string.invalid_credential_error)
-                                it.monthBirthdayState.value.isNotValidBirthMonth() -> stringResource(id = R.string.birth_month_valid_error)
-                                else -> {("")}
-                            }
-                        )
-                        BottomLineDefaultTextInput(
-                            labelResId = R.string.year,
-                            state = it,
-                            modifier = Modifier.weight(1f),
-                            transformation = VisualTransformation.None,
-                            value = it.yearBirthdayState.value,
-                            onValueChange = { state.yearBirthdayState.value = it },
                             keyboardOptions = KeyboardOptions.Default.copy( imeAction =  ImeAction.Done, keyboardType = KeyboardType.Number),
                             keyboardActions = KeyboardActions(onDone = {localFocusManager.clearFocus()}),
                             isError = when {
                                 it.isErrorRequestToFinishOutTheProfile.value -> true
-                                it.yearBirthdayState.value.isNotValidBirthYear() -> true
+                                it.weightState.value.isNotValidWeight() ->  true
                                 else -> false
                             },
                             errorMessage = when {
                                 it.isErrorRequestToFinishOutTheProfile.value -> stringResource(id = R.string.invalid_credential_error)
-                                it.yearBirthdayState.value.isNotValidBirthYear() -> stringResource(id = R.string.birth_year_valid_error)
+                                it.weightState.value.isNotValidWeight() ->  stringResource(id = R.string.weight_valid_error)
+                                else -> {("")}
+                            }
+                        )
+                        CustomDropDownMenu(
+                            modifier = Modifier
+                                .weight(1f)
+                                .fillMaxWidth(),
+                            labelResId = R.string.kicking_leg,
+                            listItems = listOf(
+                                stringResource(id = R.string.right_leg),
+                                stringResource(id = R.string.left_leg)
+                            ),
+                            value = it.workingLegState.value,
+                            onValueChange = { state.workingLegState.value = it },
+                            isError = when {
+                                it.workingLegState.value.isEmpty() -> true
+                                else -> false
+                            },
+                            errorMessage = when {
+                                it.workingLegState.value.isEmpty() -> stringResource(id = R.string.work_leg_valid_error)
                                 else -> {("")}
                             }
                         )
                     }
+                    Spacer(modifier = Modifier.size(20.dp))
+                    Text(
+                        text = stringResource(id = R.string.choose_your_playing_position),
+                        style = typography.h5,
+                        color = primaryDark
+                    )
+                    CustomDropDownMenu(
+                        labelResId = R.string.your_playing_position,
+                         modifier = Modifier.fillMaxWidth(),
+                        listItems = listOf(
+                            stringResource(id = R.string.any_position),
+                            stringResource(id = R.string.goalkeeper),
+                            stringResource(id = R.string.right_defender),
+                            stringResource(id = R.string.left_defender),
+                            stringResource(id = R.string.central_defender),
+                            stringResource(id = R.string.left_flank_defender),
+                            stringResource(id = R.string.right_flank_defender),
+                            stringResource(id = R.string.supporting_mid_defender),
+                            stringResource(id = R.string.left_mid_defender),
+                            stringResource(id = R.string.attacking_mid_defender),
+                            stringResource(id = R.string.right_winger),
+                            stringResource(id = R.string.left_winger),
+                            stringResource(id = R.string.right_flank_attacker),
+                            stringResource(id = R.string.left_flank_attacker),
+                            stringResource(id = R.string.central_forward),
+                            stringResource(id = R.string.left_forward),
+                            stringResource(id = R.string.right_forward),
+                            stringResource(id = R.string.forward_striker),
+                        ),
+                        value = it.positionState.value ,
+                        onValueChange = {state.positionState.value = it},
+                        isError = when {
+                            it.positionState.value.isEmpty() -> true
+                            else -> false
+                        },
+                        errorMessage = when {
+                            it.positionState.value.isEmpty() -> stringResource(id = R.string.position_valid_error)
+                            else -> {("")}
+                        }
+                    )
                     Spacer(modifier = Modifier.weight(1f))
                     Spacer(modifier = Modifier.size(24.dp))
                     Button(
-                        enabled = it.dayBirthdayState.value.isValidBirthDay()
-                                && it.monthBirthdayState.value.isValidBirthMonth()
-                                && it.yearBirthdayState.value.isValidBirthYear(),
-                        onClick = onFillingOutTheUserProfileStep2Clicked,
+                        enabled = it.heightState.value.isValidHeight()
+                                && it.weightState.value.isValidWeight()
+                                && it.positionState.value.isNotEmpty()
+                                && it.workingLegState.value.isNotEmpty(),
+                        onClick = onFillingOutTheUserProfileStep4Clicked,
                         modifier = Modifier
                             .fillMaxWidth()
                             .height(40.dp),
